% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storage.r
\name{ds.default.location}
\alias{ds.default.location}
\alias{ds.default.database}
\title{Default location to save data}
\usage{
ds.default.location()

ds.default.database()
}
\value{
default location to save data

default database to save data
}
\description{
The following waterfall is used to lookup the default location: 
1. options
2. environment
3. default option

Good practice is not to store this setting inside the script files. 
Add options(RTSDATA_DB='mongodb://localhost') line to the .Rprofile to use 'mongodb://localhost' database.
}
\details{
Good practice is not to store this setting inside the script files. 
Add options(RTSDATA_FOLDER='C:/Data') line to the .Rprofile to use 'C:/Data' folder.
}
\examples{
	# Default location to save data
	ds.default.location()

}
