% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{ds.get.url}
\alias{ds.get.url}
\title{Load data from URL}
\usage{
ds.get.url(
  url,
  h = curl::new_handle(),
  useragent = "Mozilla/5.0 (Windows NT 6.1; Win64; rv:62.0) Gecko/20100101",
  referer = NULL
)
}
\arguments{
\item{url}{url}

\item{h}{curl handle}

\item{useragent}{user agent}

\item{referer}{referer}
}
\description{
Load data from URL
}
\examples{
 \donttest{ 
 ds.get.url('https://finance.yahoo.com/')
 }

}
