% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{ds.load.csv}
\alias{ds.load.csv}
\title{Read csv}
\usage{
ds.load.csv(filename, sep = ",", ...)
}
\arguments{
\item{filename}{CSV filename}

\item{sep}{delimiter}

\item{...}{other parameters}
}
\description{
Read csv
}
\examples{
 # generate csv file
 filename = file.path(tempdir(), 'dummy.csv')
 cat('x1,x2,x3\n1,2,3\n', file = filename)
 ds.load.csv(filename)
 
}
