% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storage.r
\name{ds.storage.database}
\alias{ds.storage.database}
\title{MongoDB GridFS Storage model}
\usage{
ds.storage.database(url = ds.default.database(), db = "data_storage")
}
\arguments{
\item{url}{address of the mongodb server in mongo connection string URI format, 
\strong{defaults to ds.default.database database}. 

For local mongodb server, use 'mongodb://localhost' URI.
For local authenticated mongodb server, use 'mongodb://user:password@localhost' URI.}

\item{db}{name of database, \strong{defaults to 'data_storage'}}
}
\value{
list with storage options
}
\description{
MongoDB GridFS Storage model
}
\examples{
 \donttest{ 
  # change the 'yahoo' data source to use MongoDB to store historical data
  # register.data.source(src = 'yahoo', storage = ds.storage.database())
 }

}
