% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storage.r
\name{ds.storage.file.exists}
\alias{ds.storage.file.exists}
\title{Check if file exists with historical data for given ticker}
\usage{
ds.storage.file.exists(t, s)
}
\arguments{
\item{t}{ticker}

\item{s}{storage model}
}
\value{
boolean indicating if file exists with historical data for given ticker
}
\description{
Check if file exists with historical data for given ticker
}
\examples{
 ds.storage.file.exists('dummy', ds.storage.file.rdata())

}
