% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storage.r
\name{ds.storage.file.rdata}
\alias{ds.storage.file.rdata}
\title{Rdata file Storage model}
\usage{
ds.storage.file.rdata(
  location = ds.default.location(),
  extension = "Rdata",
  custom.folder = FALSE
)
}
\arguments{
\item{location}{storage location, \strong{defaults to ds.default.location folder}}

\item{extension}{file extension, \strong{defaults to 'Rdata'}}

\item{custom.folder}{custom folder flag, \strong{defaults to False}
    if flag is False \strong{default}, the data is stored at the \code{"location\src_extnsion"} folder.
if flag is True, the data is stored at the \strong{location} folder.}
}
\value{
list with storage options
}
\description{
Rdata file Storage model
}
\examples{
 # change the 'yahoo' data source to use Rdata files to store historical data
 register.data.source(src = 'yahoo', storage = ds.storage.file.rdata())

}
