% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storage.r
\name{ds.storage.file.ticker}
\alias{ds.storage.file.ticker}
\title{File with historical data for given ticker}
\usage{
ds.storage.file.ticker(t, s)
}
\arguments{
\item{t}{ticker}

\item{s}{storage model}
}
\value{
filename with historical data for given ticker
}
\description{
File with historical data for given ticker
}
\examples{
 ds.storage.file.ticker('dummy', ds.storage.file.rdata())

}
