% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functionality.r
\name{ds.functionality.default}
\alias{ds.functionality.default}
\title{Default Functionality}
\usage{
ds.functionality.default(
  check.update = TRUE,
  update.required.fn = update.required
)
}
\arguments{
\item{check.update}{flag to check for updates, \strong{defaults to TRUE}}

\item{update.required.fn}{function to check if update is required given stored historical data, 
\strong{defaults to update.required}. The update.required function takes 
last update stamp, current date/time, holiday calendar name.}
}
\value{
list with options
}
\description{
Default functionality configuration
}
\examples{
 # disable check for updates for the 'yahoo' data source
 register.data.source(src = 'yahoo', functionality = ds.functionality.default(FALSE))

}
